Add files via upload
I was able to learn how to implement Buffer Overflows and see how this vulnerability works. I got somewhat familiar with C code and learned that it has the ability to manipulate low-level details with pointers as shown throughout this entire lab. I got familiar with the term “smashing the stack” and had fun reading more about it as we were practicing hands on- running C code with no input validation, running the C code with an out-of-bounds string, and lastly, running the GDB debugger with PEDA.