Add files via upload
I was able to grasp the Linux-style access controls on files on an Ubuntu server. Some of the main objectives I really had an interest in were the impacts of setUID, getting familiar with specified permissions for groups, and customizing extended file permissions. I was able to accomplish my goal in getting familiar with an access control system and how in depth it got. In addition to that, I also enjoyed reading up more on the PAM configuration related to Linux and lastly the syslog configuration file, as they are both great resources acting as a system administrator.